refactor(dw): raname transfer page
